Changing aznamespaces over to using tsp-client. Using a branch until I can get my changes merged into tspconfig.yaml in azure-rest-api-specs.
PR here: https://github.com/Azure/azure-rest-api-specs/pull/31719
