+++ sdk/messaging/eventgrid/aznamespaces/build.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-//go:generate pwsh ./testdata/gen.ps1
+//go:generate tsp-client update
 //go:generate goimports -w .
 //go:generate go run ./internal/generate
 //go:generate goimports -w .
