--- sdk/messaging/eventgrid/aznamespaces/testdata/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "testdata",
-  "version": "0.1.0",
-  "type": "module",
-  "scripts": {
-    "pull": "pwsh ../../../../../eng/common/scripts/TypeSpec-Project-Sync.ps1 -ProjectDirectory . && rm ./TempTypeSpecFiles/Azure.Messaging.EventGrid/tspconfig.yaml",
-    "build": "tsp compile ./TempTypeSpecFiles/Azure.Messaging.EventGrid/client.tsp"
-  },
-  "dependencies": {
-    "@azure-tools/typespec-client-generator-core": "^0.42.0",
-    "@azure-tools/typespec-go": "https://github.com/Azure/autorest.go/releases/download/v0.1.0-dev.4-20240517.1/azure-tools-typespec-go-0.1.0-dev.4-20240517.1.tgz",
-    "@typespec/compiler": "^0.56.0",
-    "@typespec/versioning": "^0.56.0"
-  },
-  "overrides": {
-    "braces": "^3.0.3"
-  },
-  "private": true
-}
