--- sdk/messaging/eventgrid/aznamespaces/testdata/tsp-location.yaml
@@ -1,3 +1,0 @@
-directory: specification/eventgrid/Azure.Messaging.EventGrid
-commit: 25afa22a250c0ad5a28671003f961b9cd20d60f6
-repo: Azure/azure-rest-api-specs
